requirements.txt
commonmark==0.9.1
invoke==1.7.1
podman-compose==1.0.3
Pygments==2.13.0
python-dotenv==0.20.0
PyYAML==6.0
rich==12.5.1
toml==0.10.2
